% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additive1Distance.R
\name{additive1Distance}
\alias{additive1Distance}
\title{additive1Distance}
\usage{
additive1Distance(params, distanceMatrix, chargeMatrix)
}
\arguments{
\item{params}{A list containing the parameters for the additive dissimilarity. The entries of the list correspond to U the initial shif, v the scalar intensity of the interaction matrix and V the interaction matrix.}

\item{distanceMatrix}{The distance matrix between the unprotected attributes.}

\item{chargeMatrix}{The protected attributes as rows.}
}
\value{
A matrix giving the new distance that takes into account the original distances and the charged distances
}
\description{
Performs an additive correction to the original distance between the unprotected attributes using an interaction matrix
}
\examples{
X <- rbind(rmvnorm(50, mean = c(-1, 0.5), sigma = diag(0.25, 2)), rmvnorm(50, mean = c(-1, -0.5), sigma = diag(0.25, 2)),
           rmvnorm(50, mean = c(1, 0.5), sigma = diag(0.25, 2)), rmvnorm(50, mean = c(1, -0.5), sigma = diag(0.25, 2)))
Q <- cbind(matrix(rep(c(0, 1), 100), nrow = 2), matrix(rep(c(1, 0), 100), nrow = 2))
params = list(U = array(0, dim = c(2,2)), v = 1, V = matrix(c(1, -1, -1, 1), ncol = 2, byrow = TRUE))
add1Distance <- additive1Distance(params, as.matrix(dist(X)), t(Q))
}
\references{
E del Barrio, H Inouzhe, JM Loubes. (2019) Attraction-Repulsion clustering with applications to fairness. arXiv:1904.05254
}
