% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chargedHclust_2.R
\name{chargedHclust}
\alias{chargedHclust}
\title{chargedHclust}
\usage{
chargedHclust(X, Q, Qf, a = 0, b = 1, tipo = "single", tipoq = "additive")
}
\arguments{
\item{X}{Unprotected attribute as rows}

\item{Q}{Protected attributes as columns}

\item{Qf}{Interaction Matrix, only when tipoq = 'additive'.}

\item{a}{Intensity of perturbation, only whene tipoq in  c('additive2', multiplicative')}

\item{b}{Velocit of decrease in perturbation, only whene tipoq = 'multiplicative'}

\item{tipo}{The type of hierarchical clustering to be performet. Takes values in c('single', 'average', 'complete', 'mcquitty', 'median', 'centroid', 'ward')}

\item{tipoq}{The type of dissimilarity to use for attraction-repulsion clustering. Takes values in c('additive', 'additive2', 'multiplicative')}
}
\value{
Returns a list with elements:
\describe{
 \item{niveles}{A list with values nivel and cluster, that indicates the level at wich we are in the tree and which are the leaves to be merged at that level.}
 \item{clusters}{A list where each entry ia a partition of the data at that level of the tree.}
 \item{edges}{A graph structure for plotting the tree with igrap and ggraph.}
}
}
\description{
Performs efficient attraction-repulsion hierarchical clustring
}
\examples{
X2 <- rbind(rmvnorm(50, mean = c(-1, 0.5), sigma = diag(0.25, 2)), rmvnorm(50, mean = c(-1, -0.5), sigma = diag(0.25, 2)),
           rmvnorm(50, mean = c(1, 0.5), sigma = diag(0.25, 2)), rmvnorm(50, mean = c(1, -0.5), sigma = diag(0.25, 2)))
Q2 <- cbind(matrix(rep(c(0, 1), 100), nrow = 2), matrix(rep(c(1, 0), 100), nrow = 2))
qhcl <- chargedHclust(X2, Q2, Qf = matrix(c(0.1, 0, 0, 0.1), ncol = 2), tipo = 'centroid', tipoq = 'additive')
mygraph <- graph_from_data_frame(qhcl$edges)
ggraph(mygraph, layout = 'dendrogram', circular = FALSE) + geom_edge_diagonal() +
  geom_node_point() + theme_void() + geom_node_text(aes(label = name, filter = leaf) , angle = 90 , hjust = 1, nudge_y = -0.15) +
  ylim(-.4, NA) + ggtitle('hierarchical tree')

}
\references{
E del Barrio, H Inouzhe, JM Loubes. (2019) Attraction-Repulsion clustering with applications to fairness. arXiv:1904.05254
}
