% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e2DistProp.R
\name{e2DistProp}
\alias{e2DistProp}
\title{e2DistProp}
\usage{
e2DistProp(clusters, chargeMatrix, totalPropMatrix)
}
\arguments{
\item{clusters}{A partition.}

\item{chargeMatrix}{The protected attributes as rows.}

\item{totalPropMatrix}{A vector indicating the proportions of the protected attributes in the whole data.}
}
\value{
A list with entries:
\describe{
     \item{distance}{The average euclidean distance (over teh different clusters) between the proportions of the protected attributes in the clusters and the original proportion given by totalPropMatrix.}
     \item{clusterProportions}{The proportions of the protected attributes in each cluster.}
     \item{clusterDistance}{The average euclidean distance for each protected attribute on the different clusters with respect to original proportion given by totalPropMatrix.}
}
}
\description{
Calculates euclidean square distances between proportions of the protected atributes in the clusters of a partition and the original proportions in the whole data.
}
\examples{
X <- rbind(rmvnorm(50, mean = c(-1, 0.5), sigma = diag(0.25, 2)), rmvnorm(50, mean = c(-1, -0.5), sigma = diag(0.25, 2)),
           rmvnorm(50, mean = c(1, 0.5), sigma = diag(0.25, 2)), rmvnorm(50, mean = c(1, -0.5), sigma = diag(0.25, 2)))
Q <- cbind(matrix(rep(c(0, 1), 100), nrow = 2), matrix(rep(c(1, 0), 100), nrow = 2))
partition <- kmeans(X, 2)
e2DistProp(partition$cluster, t(Q), c(0.5,0.5))

}
