% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localDistance.R
\name{localDistance}
\alias{localDistance}
\title{localDistance}
\usage{
localDistance(params, distanceMatrix, chargeMatrix, chargedDistance)
}
\arguments{
\item{params}{A list containing the parameters for the local dissimilarity. The entries of the list correspond to u the intensity of the perturbation, v the velocity of decrease in perturbation with respect to charge, w the velocit of decrease in perturbation with respect to unperturbed distance and V the interaction matrix.}

\item{distanceMatrix}{The distance matrix between the unprotected attributes.}

\item{chargeMatrix}{The protected attributes as rows.}

\item{chargedDistance}{The euclidean distance matrix between the proteted attributes.}
}
\value{
A matrix giving the new distance that takes into account the original distances and the charged distances
}
\description{
Performs a local correction to the original distance between the unprotected attributes using an interaction matrix
}
\examples{
X <- rbind(rmvnorm(50, mean = c(-1, 0.5), sigma = diag(0.25, 2)), rmvnorm(50, mean = c(-1, -0.5), sigma = diag(0.25, 2)),
           rmvnorm(50, mean = c(1, 0.5), sigma = diag(0.25, 2)), rmvnorm(50, mean = c(1, -0.5), sigma = diag(0.25, 2)))
Q <- cbind(matrix(rep(c(0, 1), 100), nrow = 2), matrix(rep(c(1, 0), 100), nrow = 2))
params = list(u = 1, v = 1, w = 1, V = matrix(c(1, -1, -1, 1), ncol = 2, byrow = TRUE))
locDistance <- localDistance(params, as.matrix(dist(X)), t(Q), as.matrix(dist(t(Q))))
}
\references{
E del Barrio, H Inouzhe, JM Loubes. (2019) Attraction-Repulsion clustering with applications to fairness. arXiv:1904.05254
}
